Move vehicle validation to mod vehicle
It has the same structure of the backend model, where the zod validation is after the domain model.
