+++ src/app/vehicles/Vehicles.tsx
@@ -17,7 +17,7 @@ import { DataTable } from "primereact/datatable";
 import { Dialog } from "primereact/dialog";
 import { useEffect } from "react";
 import { Field, Form, FormRenderProps } from "react-final-form";
-import { emptyVehicle, Vehicle } from "./vehicle.ts";
+import { emptyVehicle, validateVehicle, Vehicle } from "./vehicle.ts";
 import { useVehicleDialog, useVehicleService } from "./vehicles.hook.ts";
 
 export function Vehicles() {
@@ -186,33 +186,6 @@ function EditVehicleForm(
         selectedVehicle,
     }: EditVehicleFormProps,
 ) {
-    const validate = (formVehicle: Vehicle) => {
-        const errors: Partial<Vehicle> = {};
-
-        if (!formVehicle.number || formVehicle.number.trim() === "") {
-            errors.number = "Vehicle number cannot be blank.";
-        }
-        else if (formVehicle.number.length > 20) {
-            errors.number = "Vehicle number maximum length is 20 characters.";
-        }
-
-        if (!formVehicle.brand || formVehicle.brand.trim() === "") {
-            errors.brand = "Vehicle brand cannot be blank.";
-        }
-        else if (formVehicle.brand.length > 100) {
-            errors.brand = "Vehicle brand maximum length is 100 characters.";
-        }
-
-        if (!formVehicle.model || formVehicle.model.trim() === "") {
-            errors.model = "Vehicle model cannot be blank.";
-        }
-        else if (formVehicle.model.length > 100) {
-            errors.model = "Vehicle model maximum length is 100 characters.";
-        }
-
-        return errors;
-    };
-
     const onSubmit = (
         formVehicle: Vehicle,
         form: FormApi<Vehicle, Vehicle>,
@@ -283,7 +256,7 @@ function EditVehicleForm(
 
     return <>
         <Form
-            validate={ validate }
+            validate={ validateVehicle }
             initialValues={ selectedVehicle ?? emptyVehicle }
             onSubmit={ onSubmit }
             render={ formBody }
